requirements.txt
alembic==1.13.2
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
ecdsa==0.19.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
gunicorn==23.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
